requirements.txt
flask
numpy
scikit-fuzzy
mysql-connector-python
